move CSS
- fix downloads
